requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
mfdata==0.3.2
pandas==2.0.0
fpdf2==2.7.0
tabula-py==2.5.0
openpyxl==3.10.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
numpy>=1.24.0
